@@ -1,6 +1,6 @@
 package io.github.gmathi.novellibrary.activity
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 
 abstract class BaseActivity : AppCompatActivity() {
