@@ -5,12 +5,12 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
-import android.support.design.widget.NavigationView
-import android.support.design.widget.Snackbar
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentTransaction
-import android.support.v4.view.GravityCompat
-import android.support.v7.widget.Toolbar
+import com.google.android.material.navigation.NavigationView
+import com.google.android.material.snackbar.Snackbar
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentTransaction
+import androidx.core.view.GravityCompat
+import androidx.appcompat.widget.Toolbar
 import android.view.MenuItem
 import android.view.ViewTreeObserver
 import co.metalab.asyncawait.async
@@ -265,18 +265,18 @@ class NavDrawerActivity : BaseActivity(), NavigationView.OnNavigationItemSelecte
     }
 
     private fun checkIntentForNotificationData() {
-        if (intent.extras != null && intent.extras.containsKey("novel")) {
-            val novel = intent.extras.getSerializable("novel") as? Novel
+        if (intent.extras != null && intent.extras!!.containsKey("novel")) {
+            val novel = intent.extras!!.getSerializable("novel") as? Novel
             novel?.let {
-                intent.extras.remove("novel")
+                intent.extras!!.remove("novel")
                 startChaptersActivity(novel)
             }
         }
     }
 
-    override fun onSaveInstanceState(outState: Bundle?) {
+    override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        outState?.putInt("currentNavId", currentNavId)
+        outState.putInt("currentNavId", currentNavId)
     }
 
     override fun onDestroy() {
