@@ -2,9 +2,9 @@ package io.github.gmathi.novellibrary.activity
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.DividerItemDecoration
+import androidx.core.content.ContextCompat
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.DividerItemDecoration
 import android.view.MenuItem
 import android.view.View
 import co.metalab.asyncawait.async
