@@ -1,8 +1,8 @@
 package io.github.gmathi.novellibrary.activity
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentTransaction
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentTransaction
 import android.view.MenuItem
 import io.github.gmathi.novellibrary.R
 import io.github.gmathi.novellibrary.fragment.SearchUrlFragment
