@@ -4,9 +4,9 @@ import android.Manifest
 import android.graphics.Rect
 import android.os.Bundle
 import android.os.Environment
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.DividerItemDecoration
-import android.support.v7.widget.RecyclerView
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.RecyclerView
 import android.text.format.Formatter
 import android.view.MenuItem
 import android.view.View
