@@ -1,8 +1,8 @@
 package io.github.gmathi.novellibrary.activity.settings
 
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.DividerItemDecoration
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.DividerItemDecoration
 import android.view.MenuItem
 import android.view.View
 import io.github.gmathi.novellibrary.R
