@@ -1,6 +1,6 @@
 package io.github.gmathi.novellibrary.adapter
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 import io.github.gmathi.novellibrary.R
