@@ -1,6 +1,6 @@
 package io.github.gmathi.novellibrary.fragment
 
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 
 
 open class BaseFragment : Fragment()
