@@ -1,7 +1,7 @@
 import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Build
-import android.support.annotation.RequiresApi
+import androidx.annotation.RequiresApi
 import android.util.Log
 import android.webkit.CookieManager
 import android.webkit.WebResourceRequest
