@@ -6,7 +6,7 @@ import android.content.res.AssetManager
 import android.graphics.Typeface
 import android.net.Uri
 import android.os.Bundle
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
