@@ -1,7 +1,7 @@
 package io.github.gmathi.novellibrary.util
 
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.helper.ItemTouchHelper
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.ItemTouchHelper
 
 class SimpleItemTouchHelperCallback(private val listener: SimpleItemTouchListener) : ItemTouchHelper.Callback() {
 
