@@ -15,10 +15,10 @@ import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.os.Environment
-import android.support.annotation.DrawableRes
-import android.support.graphics.drawable.VectorDrawableCompat
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.content.res.AppCompatResources
+import androidx.annotation.DrawableRes
+import androidx.vectordrawable.graphics.drawable.VectorDrawableCompat
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.content.res.AppCompatResources
 import android.util.TypedValue
 import com.afollestad.materialdialogs.MaterialDialog
 import io.github.gmathi.novellibrary.BuildConfig
@@ -140,7 +140,7 @@ object Utils {
             inChannel!!.transferTo(0, inChannel.size(), outChannel)
         } finally {
             inChannel?.close()
-            outChannel?.close()
+            outChannel.close()
         }
     }
 
@@ -174,13 +174,13 @@ object Utils {
         if (dir.exists()) {
             var result: Long = 0
             val fileList = dir.listFiles()
-            for (i in fileList.indices) {
+            for (i in fileList!!.indices) {
                 // Recursive call if it's a directory
-                if (fileList[i].isDirectory) {
-                    result += getFolderSize(fileList[i])
+                result += if (fileList[i].isDirectory) {
+                    getFolderSize(fileList[i])
                 } else {
                     // Sum the file size in bytes
-                    result += fileList[i].length()
+                    fileList[i].length()
                 }
             }
             return result // return the file size
@@ -229,15 +229,15 @@ object Utils {
     fun getBitmapFromDrawable(context: Context, @DrawableRes drawableId: Int): Bitmap {
         val drawable = AppCompatResources.getDrawable(context, drawableId)
 
-        if (drawable is BitmapDrawable) {
-            return drawable.bitmap
+        return if (drawable is BitmapDrawable) {
+            drawable.bitmap
         } else if (drawable is VectorDrawableCompat || drawable is VectorDrawable) {
             val bitmap = Bitmap.createBitmap(drawable.intrinsicWidth, drawable.intrinsicHeight, Bitmap.Config.ARGB_8888)
             val canvas = Canvas(bitmap)
             drawable.setBounds(0, 0, canvas.width, canvas.height)
             drawable.draw(canvas)
 
-            return bitmap
+            bitmap
         } else {
             throw IllegalArgumentException("unsupported drawable type")
         }
